accepted solution from https://sourceforge.net/tracker/?func=detail&atid=879320&aid=3582197&group_id=176959
If the power of the receive signal is zero (as set by the analogue model filtering, e.g. Radio State when the radio is not in receive), returning a BER of zero leads to a successful SFD detection and abusive "Start receiving" control message sent to the MAC. This is wrong because the SFD of a null signal should not be detected.

Therefore I suggest to return 1.0 instead of 0.0. I am aware that a BER of 1 is meaningless theoretically, but at least it will lead to a correct program flow control as only this value ensures that all packets received with a zero power are not detected.